Run the unit tests on both i386 and x86_64 The test harness library is now built as a universal binary, and the tests are run on both architectures using the arch command.
Some changes to the Objective-C test files were necessary.
On i386, the "fragile" Objective-C ABI is used by default, which
requires listing ivars explicitly in the @interface declaration and
using @synthesize to generate getters and setters.
clang has a -fobjc-nonfragile-abi flag to force using the new,
non-fragile ABI, but for me this caused linker errors when linking the
i386 part of the object files.