Fix architecture detection and add test cases for send_message The architecture returned by platform.machine() indicates the machine's physical architecture - this means that 32-bit Python on x86_64 gets "x86_64" back from platform.machine().
The reason why this didn't cause issues before is that directly calling
a method on an object uses a different mechanism than send_message does.
When calling a method on an object, the ObjCInstance internally looks up
the method's implementation and calls it with the correct signature,
without ever using objc_msgSend and friends.
send_message is almost never used, but does use objc_msgSend, and it
previously handled selection of the objc_msgSend variants incorrectly.

Now there are test cases for both ways of calling methods, so hopefully
things should work now.