Fix ObjCDictInstance.copy
It's not always possible to call dictionaryWithDictionary: on the
dictionary's runtime class. Because NSDictionary is a class cluster,
NSDictionary objects are actually instances of private subclasses,
not of NSDictionary itself. These private subclasses do not always have
a working dictionaryWithDictionary: class method. So we call it on
NSMutableDictionary instead, where it is guaranteed to exist and work.