Remove a bad type test in test_NSDictionary
The dictionary returned from dictionaryWithDictionary: is not required
to be immutable, even when calling it on NSDictionary and not
NSMutableDictionary. The returned dictionary can also be immutable
even though its class is a NSMutableDictionary subclass. (This is the
case for __NSCFDictionary for example.) In that case, attempting to
mutate it causes an Objective-C runtime exception and not a Python
TypeError.

At the moment our copy method always calls NSMutableDictionary's
dictionaryWithDictionary:, which makes the test *always* fail, but due
to the reasons explained above, the test would not be guaranteed to
work even if it would call NSDictionary's dictionaryWithDictionary:
instead.