Update Sphinx Makefile and make.bat
Regenerated the files using Sphinx version 2.4.4's sphinx-quickstart.
They are now much shorter than before, because most of the logic is now
implemented as part of sphinx-build instead of pure Make/Batch.

The previous version of the Makefile contained some custom commands
related to virtual environments. I've removed them, for multiple
reasons:

* the use of the virtual environment wasn't consistent (some of the
  Makefile targets automatically activated the virtual environment,
  others didn't)
* they don't seem to be mentioned anywhere in the docs
* they were only available in the Makefile and not the Windows make.bat
* they assume a specific virtual environment tool (virtualenv) and
  location (env), which may not be what the user wants to use

The user is now expected to manually create and activate a virtual
environment, as described in the tutorial.

The Makefile target for sphinx-autobuild has been renamed from run to
livehtml (based on
https://github.com/GaretJax/sphinx-autobuild/blob/e0f40b6359ca00a48ca5ba903de1e25da27de453/README.rst#makefile-integration
). The behavior has also changed slightly - it no longer automatically
activates a virtual environment (see above), and it uses the same build
directory as all other targets (_build) instead of its own directory
(build). The livehtml target is also available in the Windows make.bat.