Set package metadata declaratively in setup.cfg where possible
This significantly shortens the code in setup.py. The only parts of the
metadata that still need to be set programmatically in setup.py are the
version (which needs to be parsed from rubicon/objc/__init__.py using a
regex) and test_suite (which is only used by the deprecated
setup.py test command and cannot be set in setup.cfg).