Remove bit fields from "struct complex" test structure
ctypes (and the underlying libffi library) does not support passing
structures containing bit fields by value. In the past this has
silently worked without issues on x86 and x86_64, despite not being
officially supported. However, Python 3.7.6 and Python 3.8.1 added an
explicit check in ctypes to disallow this. This check has been removed
again in Python 3.7.7 and Python 3.8.2 to avoid breaking working code
in a bugfix release of Python, but Python 3.9 will probably contain the
check again. In any case, it's safer to not rely on unsupported
behavior like this in our tests.

Relevant CPython and libffi issues:

* https://bugs.python.org/issue39295
* https://bugs.python.org/issue16576
* https://bugs.python.org/issue16575
* libffi/libffi#33