Move Makefile into tests/objc
This Makefile is only used to build the test harness. It was previously
kept in the root of the repo so that it can be used more conveniently
(make instead of make -C tests/objc), but now that make is run
automatically by our tox configuration, this is no longer an issue.