Disallow passing class name strings into send_message
This feature is rarely used - it had only a single usage in the
rubicon-objc codebase, and is probably even less used outside, because
most code can use high-level ObjCClass method syntax instead of
send_message.

The few cases that used this behavior can be easily updated to use an
explicit get_class or ObjCClass call instead, so there's not much
benefit in having this special case in send_message.