Change send_message vararg syntax and check number of arguments
Previously send_message accepted more arguments than are specified in
argtypes. This is useful when calling variadic methods, but is easy to
misuse by accident, because extra arguments passed to non-variadic
methods are silently ignored. Varargs now need to be passed as a list
into a separate kwarg, which allows checking the number of non-variadic
args against the number of argtypes.