Remove automatic Python conversion of NS[Decimal]Number for performance
Improves the bad method call performance described in #183.

This roughly doubles the speed of calls to simple Objective-C methods
that return an object, because the returned object no longer has two
messages sent to it automatically.

This is a breaking change, but a necessary one, because of the
significant performance impact of the automatic conversion. Most code
shouldn't be affected by this change though.

This change also allows removing the private _auto kwarg from
py_from_ns, which makes that function's logic a bit easier to
understand.