Merge pre_register and register methods used internally by ObjCClass
The new method that replaces the two is called class_register, but
behaves exactly like pre_register did previously - it receives a Class
pointer rather than an ObjCClass object and is called before the
Objective-C class is registered. The name change is for symmetry with
protocol_register and to avoid confusion with the previous names.

This simplifies the code in ObjCClass.__new__ - now only one
registration method needs to be called, in a single place in
_new_from_class_statement. Because now all registration methods are
called before the ObjCClass object is created, there is also no more
need to reload the class's methods.